s02e04/requirements.txt
openai>=1.90.0
Pillow>=11
requests>=2.32.4
python-dotenv>=1.1.0